themes [wip]
added the code for changing the way trains, powerups, and coins look depending on the current theme.

still need to add code for the buttons to change appearences, but it should be really simple.
oh and also there isnt a menu to buy or select themes at all, i plan to use the same dynamic menu that the lettuce dealer has. speaking of the dynamic shop menu, it will be themed for the respective store, rather than whatever theme you buy. for example, the lettuce dealer one will be theme after the lettuce dealer.