lettuce customs/theme wip
added the menus for buying themes, and switching to them.

all thats left now (for themes) is to actually create and import the textures for the new themes.

and also make the a way to change the soundtrack. the idea is a shop called like lettuce jams, the charachter for the shop is lettuce but in a jelly jar (how creative ik) or maybe a boombox with lettuce speakers (also very creative) and he sells custom music for the game. ill probably maybe add the OG music as an option though that may be difficult to do as one entire thing as in the old game music was tied to what skin you had, so setting music to og would require a lot of checks to see what skin you are and change the music to that skin, then again that would be problamatic for someone who wants the chad lettuce music enabled when they're not using the skin and what not.

hmmm, yeah thats probably not the best way to implement to og music.

it seems like the best way to do that is have all the og music bought at once, but you can change what run music it plays in the settings menu.